Quite a lot of changes ! This commit alone won't build as is as it will require the next upcoming commits to build.
Moving Drawing function to a .C file as to avoid issues with types.
Adding a scanline mode as to mimic the real hardware behaviour.
For now, it is not configurable.

Some changes had to be done to SDL surfaces too.

There's an issue with some games rotating to the right (Klax and Gauntlet)
with 8 black pixels on top that needs to be cropped for proper look.
Hence why we need to use Cut_Off_Y for this purpose, it stays 0 otherwise.
This should fix those games.